[refactor] WebLinkScreen: create UIEvent and move UIState up from viewmodel package
