+++ app/src/main/java/uk/ryanwong/skycatnews/weblink/ui/WebLinkUIEvent.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2023. Ryan Wong (hello@ryanwong.co.uk)
+ */
+
+package uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.weblink.ui
+
+data class WebLinkUIEvent(
+    val onErrorShown: (errorId: Long) -> Unit,
+)
