+++ app/src/main/java/uk/ryanwong/skycatnews/weblink/ui/WebLinkUIState.kt
@@ -2,7 +2,7 @@
  * Copyright (c) 2022. Ryan Wong (hello@ryanwong.co.uk)
  */
 
-package uk.ryanwong.skycatnews.weblink.ui.viewmodel
+package uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.weblink.ui
 
 import uk.ryanwong.skycatnews.app.util.ErrorMessage
