+++ app/src/main/java/uk/ryanwong/skycatnews/weblink/ui/screen/WebLinkScreen.kt
@@ -21,13 +21,14 @@ import androidx.compose.ui.semantics.semantics
 import com.google.accompanist.web.WebView
 import com.google.accompanist.web.rememberWebViewState
 import uk.ryanwong.skycatnews.R
-import uk.ryanwong.skycatnews.weblink.ui.viewmodel.WebLinkUIState
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.weblink.ui.WebLinkUIEvent
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.weblink.ui.WebLinkUIState
 
 @Composable
 fun WebLinkScreen(
     modifier: Modifier = Modifier,
     uiState: WebLinkUIState,
-    onErrorShown: (errorId: Long) -> Unit,
+    uiEvent: WebLinkUIEvent,
 ) {
     val snackbarHostState = remember { SnackbarHostState() }
 
@@ -59,7 +60,7 @@ fun WebLinkScreen(
                 message = errorMessageText,
                 actionLabel = actionLabel
             )
-            onErrorShown(errorMessage.id)
+            uiEvent.onErrorShown(errorMessage.id)
         }
     }
 }
