+++ app/src/main/java/uk/ryanwong/skycatnews/weblink/ui/viewmodel/WebLinkViewModel.kt
@@ -20,6 +20,7 @@ import uk.ryanwong.skycatnews.app.di.DispatcherModule
 import uk.ryanwong.skycatnews.app.util.ErrorMessage
 import uk.ryanwong.skycatnews.newslist.data.repository.NewsListRepository
 import uk.ryanwong.skycatnews.newslist.domain.model.NewsItem
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.weblink.ui.WebLinkUIState
 
 @HiltViewModel
 class WebLinkViewModel @Inject constructor(
