+++ app/src/test/java/uk/ryanwong/skycatnews/newslist/ui/viewmodel/WebLinkViewModelTest.kt
@@ -18,7 +18,7 @@ import kotlinx.coroutines.test.UnconfinedTestDispatcher
 import kotlinx.coroutines.test.runTest
 import uk.ryanwong.skycatnews.R
 import uk.ryanwong.skycatnews.newslist.data.repository.MockNewsListRepository
-import uk.ryanwong.skycatnews.weblink.ui.viewmodel.WebLinkUIState
+import uk.ryanwong.skycatnews.uk.ryanwong.skycatnews.weblink.ui.WebLinkUIState
 import uk.ryanwong.skycatnews.weblink.ui.viewmodel.WebLinkViewModel
 
 @OptIn(ExperimentalCoroutinesApi::class)
